+++ demo_proj/demo_app/views.py
@@ -1,3 +1,11 @@
-from django.shortcuts import render
+from django.views.generic import TemplateView
+
 
 # Create your views here.
+
+class HomeView(TemplateView):
+    template_name = "home.html"
+
+
+class Dashboard(TemplateView):
+    template_name = "dashboard.html"