+++ slick_reporting/views.py
@@ -44,9 +44,7 @@ class ExportToCSV(object):
 
     def get_response(self):
         response = HttpResponse(content_type="text/csv")
-        response["Content-Disposition"] = "attachment; filename={filename}.csv".format(
-            filename=self.get_filename()
-        )
+        response["Content-Disposition"] = "attachment; filename={filename}.csv".format(filename=self.get_filename())
 
         writer = csv.writer(response)
         for rows in self.get_rows():
@@ -61,9 +59,7 @@ class ExportToCSV(object):
             yield [line[col_name] for col_name in columns]
 
     def get_columns(self, extra_context=None):
-        return list(
-            zip(*[(x["name"], x["verbose_name"]) for x in self.report_data["columns"]])
-        )
+        return list(zip(*[(x["name"], x["verbose_name"]) for x in self.report_data["columns"]]))
 
     def __init__(self, request, report_data, report_title, **kwargs):
         self.request = request
@@ -85,9 +81,7 @@ class ExportToStreamingCSV(ExportToCSV):
             (writer.writerow(row) for row in self.get_rows()),
             content_type="text/csv",
             headers={
-                "Content-Disposition": 'attachment; filename="{filename}.csv"'.format(
-                    filename=self.get_filename()
-                )
+                "Content-Disposition": 'attachment; filename="{filename}.csv"'.format(filename=self.get_filename())
             },
         )
 
@@ -97,7 +91,7 @@ class ReportViewBase(ReportGeneratorAPI, FormView):
 
     report_title = ""
 
-    report_title_context_key = "title"
+    report_title_context_key = "report_title"
 
     report_generator_class = ReportGenerator
 
@@ -118,15 +112,26 @@ class ReportViewBase(ReportGeneratorAPI, FormView):
     doc_type_field_name = "doc_type"
     doc_type_plus_list = None
     doc_type_minus_list = None
+    auto_load = True
 
     default_order_by = ""
 
-    template_name = "slick_reporting/simple_report.html"
+    template_name = "slick_reporting/report.html"
 
     @staticmethod
     def form_filter_func(fkeys_dict):
         # todo revise
         return fkeys_dict
+
+    @classmethod
+    def get_report_title(cls):
+        """
+        :return: The report name
+        """
+        name = cls.__name__
+        if cls.report_title:
+            name = cls.report_title
+        return name
 
     def order_results(self, data):
         """
@@ -134,9 +139,7 @@ class ReportViewBase(ReportGeneratorAPI, FormView):
         :param data: List of Dict to be ordered
         :return: Ordered data
         """
-        order_field, asc = OrderByForm(self.request.GET).get_order_by(
-            self.default_order_by
-        )
+        order_field, asc = OrderByForm(self.request.GET).get_order_by(self.default_order_by)
         if order_field:
             data = dictsort(data, order_field, asc)
         return data
@@ -144,13 +147,9 @@ class ReportViewBase(ReportGeneratorAPI, FormView):
     def get_doc_types_q_filters(self):
         if self.doc_type_plus_list or self.doc_type_minus_list:
             return (
-                [Q(**{f"{self.doc_type_field_name}__in": self.doc_type_plus_list})]
-                if self.doc_type_plus_list
-                else []
+                [Q(**{f"{self.doc_type_field_name}__in": self.doc_type_plus_list})] if self.doc_type_plus_list else []
             ), (
-                [Q(**{f"{self.doc_type_field_name}__in": self.doc_type_minus_list})]
-                if self.doc_type_minus_list
-                else []
+                [Q(**{f"{self.doc_type_field_name}__in": self.doc_type_minus_list})] if self.doc_type_minus_list else []
             )
 
         return [], []
@@ -158,31 +157,31 @@ class ReportViewBase(ReportGeneratorAPI, FormView):
     def get(self, request, *args, **kwargs):
         form_class = self.get_form_class()
         self.form = self.get_form(form_class)
+        report_data = {}
         if self.form.is_valid():
-            report_data = self.get_report_results()
+            if self.request.GET or self.request.POST or request.headers.get("x-requested-with") == "XMLHttpRequest":
+                # only display results if it's requested,
+                # considered requested if it's ajax request, or a populated GET or POST.
+                report_data = self.get_report_results()
 
-            export_option = request.GET.get("_export", "")
-            if export_option:
-                try:
-                    return getattr(self, f"export_{export_option}")(report_data)
-                except AttributeError:
-                    pass
+                export_option = request.GET.get("_export", "")
+                if export_option:
+                    try:
+                        return getattr(self, f"export_{export_option}")(report_data)
+                    except AttributeError:
+                        pass
 
-            if request.headers.get("x-requested-with") == "XMLHttpRequest":
-                return self.ajax_render_to_response(report_data)
+                if request.headers.get("x-requested-with") == "XMLHttpRequest":
+                    return self.ajax_render_to_response(report_data)
 
-            return self.render_to_response(
-                self.get_context_data(report_data=report_data)
-            )
+            return self.render_to_response(self.get_context_data(report_data=report_data))
         else:
             return self.form_invalid(self.form)
 
         # return self.render_to_response(self.get_context_data())
 
     def export_csv(self, report_data):
-        return self.csv_export_class(
-            self.request, report_data, self.report_title
-        ).get_response()
+        return self.csv_export_class(self.request, report_data, self.report_title).get_response()
 
     @classmethod
     def get_report_model(cls):
@@ -191,9 +190,7 @@ class ReportViewBase(ReportGeneratorAPI, FormView):
         return cls.report_model
 
     def ajax_render_to_response(self, report_data):
-        return HttpResponse(
-            self.serialize_to_json(report_data), content_type="application/json"
-        )
+        return HttpResponse(self.serialize_to_json(report_data), content_type="application/json")
 
     def serialize_to_json(self, response_data):
         """Returns the JSON string for the compiled data object."""
@@ -210,9 +207,7 @@ class ReportViewBase(ReportGeneratorAPI, FormView):
         if settings.DEBUG:
             indent = 4
 
-        return json.dumps(
-            response_data, indent=indent, use_decimal=True, default=date_handler
-        )
+        return json.dumps(response_data, indent=indent, use_decimal=True, default=date_handler)
 
     def get_form_class(self):
         """
@@ -257,11 +252,18 @@ class ReportViewBase(ReportGeneratorAPI, FormView):
             )
         return kwargs
 
+    def get_crosstab_ids(self):
+        """
+        Hook to get the crosstab ids
+        :return:
+        """
+        return self.form.get_crosstab_ids()
+
     def get_report_generator(self, queryset, for_print):
         q_filters, kw_filters = self.form.get_filters()
         crosstab_compute_remainder = False
         if self.crosstab_field:
-            self.crosstab_ids = self.form.get_crosstab_ids()
+            self.crosstab_ids = self.get_crosstab_ids()
         try:
             crosstab_compute_remainder = (
                 self.form.get_crosstab_compute_remainder()
@@ -302,6 +304,7 @@ class ReportViewBase(ReportGeneratorAPI, FormView):
             crosstab_ids=self.crosstab_ids,
             crosstab_columns=self.crosstab_columns,
             crosstab_compute_remainder=crosstab_compute_remainder,
+            crosstab_ids_custom_filters=self.crosstab_ids_custom_filters,
             format_row_func=self.format_row,
             container_class=self,
             doc_type_plus_list=doc_type_plus_list,
@@ -356,9 +359,7 @@ class ReportViewBase(ReportGeneratorAPI, FormView):
         """
         Ensure the sane settings are passed to the front end.
         """
-        return generator.get_chart_settings(
-            self.chart_settings or [], self.report_title
-        )
+        return generator.get_chart_settings(self.chart_settings or [], self.report_title)
 
     @classmethod
     def get_queryset(cls):
@@ -401,6 +402,7 @@ class ReportViewBase(ReportGeneratorAPI, FormView):
         context = super().get_context_data(**kwargs)
         context[self.report_title_context_key] = self.report_title
         context["crispy_helper"] = self.get_form_crispy_helper()
+        context["auto_load"] = self.auto_load
 
         if not (self.request.POST or self.request.GET):
             # initialize empty form with initials if the no data is in the get or the post
@@ -478,6 +480,8 @@ class SlickReportingListViewMixin:
 
         return self.report_generator_class(
             self.get_report_model(),
+            start_date=self.form.get_start_date(),
+            end_date=self.form.get_end_date(),
             q_filters=q_filters,
             kwargs_filters=kw_filters,
             date_field=self.date_field,